Misc. controller/route clean up
