+++ src/controllers/index.ts
@@ -1,0 +1,2 @@
+export * from "./userControllers";
+export * from "./savedChannelController";
