+++ src/routes/userRoutes.ts
@@ -1,20 +1,20 @@
+/**
+ * API routes.
+ * Base path /api/users
+ */
+
 import {
-  getSavedChannelsController,
   addSavedChannelController,
-  deleteSavedChannelController,
   addUserController,
+  deleteSavedChannelController,
   deleteUserController,
-} from "@/controllers/userControllers";
+  getSavedChannelsController,
+} from "@/controllers";
 import { checkJwt } from "@/middleware/checkJwtMiddleware";
 import express from "express";
 
 const router = express.Router();
 
-// Base path /api/users
-
-// TODO: Create controllers for these routes
-// User should send requests with appropriate access token, enabling us to access routes
-// For user delete, we will likely need to get a management API access token. Use deleteAuth0User function from `resports-aws-cdk` for this if needed
 router
   .route("/")
   .get(checkJwt, addUserController)
